requirements.txt
streamlit>=1.49.0
supabase>=2.18.0
st-supabase-connection>=0.1.0
pillow>=10.0.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.17.0

